Move settings and auth to cache
