+++ app/manage/lib/core/cache/settings.dart
