+++ app/manage/lib/core/router/manage_route_information_parser.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:manage/core/auth.dart';
+import 'package:manage/core/cache/auth.dart';
 import 'package:manage/core/router/manage_route_path.dart';
 
 class ManageRouteInformationParser
