+++ app/manage/lib/main.dart
@@ -4,11 +4,11 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart' show WidgetsFlutterBinding, runApp;
 import 'package:hive/hive.dart';
 import 'package:hive_flutter/hive_flutter.dart';
-import 'package:manage/core/auth.dart';
+import 'package:manage/core/cache/auth.dart';
 import 'package:path_provider/path_provider.dart';
 
 import 'core/app.dart';
-import 'core/settings.dart';
+import 'core/cache/settings.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
