+++ app/manage/test/theme_mode_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:manage/core/settings.dart';
+import 'package:manage/core/cache/settings.dart';
 import 'package:manage/core/theme.dart';
 import 'package:flutter/scheduler.dart';
